Refactor and update unit tests
Refactor the unit tests into separate files for the different datasets,
and remove some arguments specifying constant-valued parameters from the
calls to CDSDataStore.open_data in unit tests, in anticipation of
removal of these parameters from the open parameter schemas.
